// frontend/src/client/models/MentorServiceCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MentorServiceCreate = {
    mentor_id: number;
    title: string;
    description?: (string | null);
    banner_url?: (string | null);
    price_usd?: (number | null);
    estimated_duration_minutes?: (number | null);
    category?: (string | null);
    highlights?: (Array<string> | null);
};