// frontend/src/client/models/MentorServiceUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MentorServiceUpdate = {
    title?: (string | null);
    description?: (string | null);
    banner_url?: (string | null);
    price_usd?: (number | null);
    estimated_duration_minutes?: (number | null);
    category?: (string | null);
    highlights?: (Array<string> | null);
    is_active?: (boolean | null);
};